Increased test coverage in main package
